WoL 기능 추가, 엔티티 및 DTO 자료형 수정, 뷰 구현, 엔티티 구조 변경, 버그 수정
- API에 내부망 WoL기능 추가
- 모든 엔티티 자료형 클래스로 수정
- 트리거 생성 뷰 구현
- 트리거 액션 엔티티 수정
- 트리거 등록기능 구현
- 트리거 리스트 출력기능 구현
- 신규 노드 등록요청시 enableStatus 컬럼 값 없어서 등록 안되는 버그 수정
등등..

+++ src/main/java/net/softbell/bsh/domain/entity/NodeItemHistory.java
@@ -42,10 +42,10 @@ public class NodeItemHistory implements Serializable
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="item_history_id", unique=true, nullable=false)
-	private long itemHistoryId;
+	private Long itemHistoryId;
 
 	@Column(name="item_status", nullable=false)
-	private long itemStatus;
+	private Double itemStatus;
 
 	@Temporal(TemporalType.TIMESTAMP)
 	@Column(name="receive_date", nullable=false)
